% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_config.R
\name{add_missing_function_fields}
\alias{add_missing_function_fields}
\title{Add any missing required fields to the config using defaults where available.}
\usage{
add_missing_function_fields(info, fields)
}
\arguments{
\item{info}{The config loaded from file.}

\item{fields}{Collection of fields to set defaults for.}
}
\description{
This defaults to using the field name for the function name.
For the 'test' field also sets the verification queries function default and
the test file default.
}
\keyword{internal}
