% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_config.R
\name{add_missing_required_fields}
\alias{add_missing_required_fields}
\title{Add any missing required fields to the config.}
\usage{
add_missing_required_fields(info, required)
}
\arguments{
\item{info}{The config loaded from file.}

\item{required}{Collection of required fields.}
}
\description{
This defaults to using the field name for the function and file name equal
to the field, and function name. For the 'test' field only need the file.
}
\keyword{internal}
