% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_import.R
\name{build_log_data}
\alias{build_log_data}
\title{Build the data to be written to the import log}
\usage{
build_log_data(import_path, comment)
}
\arguments{
\item{import_path}{Path to the import process directory.}

\item{comment}{An optional comment for this import run.}
}
\value{
The prepared data for the log table. This is the name
of the import, the date, comment and git information including
user name, user email, current branch and hash of HEAD.
}
\description{
Build the data to be written to the import log
}
\keyword{internal}
