% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into_returning.R
\name{convert_bit64_columns}
\alias{convert_bit64_columns}
\title{Convert any bit64 type columns to integers.}
\usage{
convert_bit64_columns(data)
}
\arguments{
\item{data}{The data frame with columns to be converted.}
}
\value{
Data frame with with any bit64 columns converted to integers.
}
\description{
Postgres returns bit64 types for integer fields.
}
\keyword{internal}
