% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl.R
\name{DataImport}
\alias{DataImport}
\title{Data Import}
\description{
Manage data import.
}
\section{Methods}{


\describe{
\item{\code{extract}}{
Run the extract stage of the data import
\cr\emph{Usage:}\preformatted{extract()}
}
\item{\code{transform}}{
Run the transform stage of the data import
\cr\emph{Usage:}\preformatted{transform()}
}
\item{\code{load}}{
Run the load stage of the data import
\cr\emph{Usage:}\preformatted{load(comment = NULL, dry_run = FALSE, force = FALSE)}

\emph{Arguments:}
\itemize{
\item{\code{comment}:   Optional comment which will be persisted in the log of the import run in the database.
}\preformatted{\item{\code{dry_run}:   Whether to run in dry run mode. If TRUE then any database changes will be rolled back. Defaults to FALSE.
}

\item{\code{force}:   If TRUE then checks that repo is up to date with git remote will be skipped. Defautls to FALSE.
}
}

}
}
\item{\code{get_extracted_data}}{
Get the extracted data created by the extract step
\cr\emph{Usage:}\preformatted{get_extracted_data()}
}
\item{\code{get_transformed_data}}{
Get the transformed data created by the transform step
\cr\emph{Usage:}\preformatted{get_transformed_data()}
}
}
}

\examples{
path <- dettl:::prepare_test_import(
  system.file("examples", "person_information", package = "dettl"),
  system.file("examples", "dettl_config.yml", package = "dettl"))
import_path <- file.path(path, "person_information")

import <- dettl::dettl(import_path, db_name = "test")
import$extract()
import$transform()
import$load()
}
