% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_config.R
\name{db_config}
\alias{db_config}
\title{Get db config from path.}
\usage{
db_config(path)
}
\arguments{
\item{path}{Path to directory containing path.}
}
\value{
The config as a db_config object.
}
\description{
This locates config from dettl_config.yml containing database connection
configuration.
}
\keyword{internal}
