% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_config.R
\name{db_config_read_yaml}
\alias{db_config_read_yaml}
\title{Read yaml file representing a db config.}
\usage{
db_config_read_yaml(filename, path)
}
\arguments{
\item{filename}{Path with filename to config yaml file.}

\item{path}{Path to directory containing path.}
}
\value{
The config as a dettl_config object.
}
\description{
Reads the file and parses the returned data.
}
\keyword{internal}
