% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect}
\alias{db_connect}
\title{Connect to the database configured via yaml.}
\usage{
db_connect(type, path)
}
\arguments{
\item{type}{The db to connect to, must match a db configured in db config.}

\item{path}{Path to directory contianing yaml config.}
}
\description{
Connect to the database configured via yaml.
}
\keyword{internal}
