% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_get_log_table}
\alias{db_get_log_table}
\title{Get the log table name from configuration for a particular db}
\usage{
db_get_log_table(type, path)
}
\arguments{
\item{type}{The db to get the log table for, must match a db configured in
db config.}

\item{path}{Path to directory contianing yaml config.}
}
\description{
Get the log table name from configuration for a particular db
}
\keyword{internal}
