% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbi_db_connect}
\alias{dbi_db_connect}
\title{Create a connection a DBMS.}
\usage{
dbi_db_connect(drv, ...)
}
\description{
Uses \code{\link[DBI]{dbConnect}} to connect to a DBMS. If this uses
\code{\link[RSQLite]{SQLite}} then it will ensure foreign key constraints are
enabled.
}
\keyword{internal}
