% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_auto_load_create.R
\name{dettl_auto_load_create}
\alias{dettl_auto_load_create}
\title{Automatic load function for create mode imports.}
\usage{
dettl_auto_load_create(transformed_data, con)
}
\arguments{
\item{transformed_data}{The list of transformed data frames to create in the
database.}

\item{con}{Connection to the database to add data to.}
}
\description{
The automatic load function loops over the transformed data and creates
a table from each data frame. It uses the default types in the db from the
types in the data frame and then appends the rows to the newly created
tables.
}
\details{
This will now allow appending to an existing table.
}
\keyword{internal}
