% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_config.R
\name{dettl_config}
\alias{dettl_config}
\title{Get db config from path.}
\usage{
dettl_config(path)
}
\arguments{
\item{path}{Path to directory containing path.}
}
\value{
The config as a dettl_config object.
}
\description{
This locates config from dettl_config.yml containing database connection
configuration.
}
\keyword{internal}
