% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{dettl_create_log_table}
\alias{dettl_create_log_table}
\title{Initialise the database by creating log table if it doesn't already exist}
\usage{
dettl_create_log_table(path, db_name)
}
\arguments{
\item{path}{Path to import directory containing db connection configuration.}

\item{db_name}{The name of the db to connect to. Connection info must be
configured via the \code{dettl_config.yml}.}
}
\description{
Initialise the database by creating log table if it doesn't already exist
}
\examples{
path <- dettl:::prepare_test_import(
  system.file("examples", "person_information", package = "dettl"),
  system.file("examples", "dettl_config.yml", package = "dettl"),
  add_log_table = FALSE
)
dettl::dettl_create_log_table(file.path(path, "person_information"), "test")
}
