% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_config.R
\name{dettl_locate_config}
\alias{dettl_locate_config}
\title{Locate the directory containing the config file.}
\usage{
dettl_locate_config(path)
}
\arguments{
\item{path}{Path to directory to start looking in.}
}
\value{
The directory containing the dettl_config yaml file.
}
\description{
Locate the directory containing the config file.
}
\keyword{internal}
