% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_runner.R
\name{dettl_run_extract}
\alias{dettl_run_extract}
\title{Run extract stage of an import}
\usage{
dettl_run_extract(path, db_name = NULL)
}
\arguments{
\item{path}{Path to the import directory.}

\item{db_name}{The name of the db to connect to. Connection info must be
configured via the `dettl_config.yml`. If name is left blank this will default
to using the first db configured.}
}
\value{
The extracted data.
}
\description{
Run extract stage of an import
}
\examples{
path <- dettl:::prepare_test_import(
  dettl:::dettl_file("examples", "person_information"),
  dettl:::dettl_file("examples", "dettl_config.yml")
)
dettl::dettl_run_extract(file.path(path, "person_information/"))
}
