% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_runner.R
\name{dettl_run_load}
\alias{dettl_run_load}
\title{Run up to and including the load stage of an import}
\usage{
dettl_run_load(path, db_name = NULL, comment = NULL, dry_run = FALSE,
  force = FALSE)
}
\arguments{
\item{path}{Path to the import directory.}

\item{db_name}{The name of the db to connect to. Connection info must be
configured via the `dettl_config.yml`. If name is left blank this will default
to using the first db configured.}

\item{comment}{Optional comment to be written to db log table when import is
run.}

\item{dry_run}{If TRUE then any changes to the database will be rolled back.}

\item{force}{If TRUE then checks the import is up to date with remote git
repo will be skipped.}
}
\description{
This will run all stages and make updates to the database unless run in
dry_run mode.
}
\examples{
path <- dettl::prepare_test_import(
  system.file("examples", "person_information", package = "dettl"),
  system.file("examples", "dettl_config.yml", package = "dettl")
)
dettl::dettl_run_load(file.path(path, "person_information/"), "test",
  comment = "Example import")
}
