% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{dettl_save}
\alias{dettl_save}
\title{Save data}
\usage{
dettl_save(import, file, stage)
}
\arguments{
\item{import}{The import object to save the data for.}

\item{file}{File path at which to save the data.}

\item{stage}{The stage or stages to save. 'extract' and/or 'transform'}
}
\description{
Saves any extracted and/or transformed data as separate sheet of an xlsx
file.
}
\examples{
path <- dettl:::prepare_test_import(
  system.file("examples", "person_information", package = "dettl"),
  system.file("examples", "dettl_config.yml", package = "dettl")
)
import <- dettl(file.path(path, "person_information"), "test")
run_import(import, c("extract", "transform"), dry_run = TRUE)
t <- tempfile()
dettl_save(import, t, "extract")
t2 <- tempfile()
dettl_save(import, t2, "transform")
t3 <- tempfile()
dettl_save(import, t3, c("extract", "transform"))
}
