% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{expand_wildcards}
\alias{expand_wildcards}
\title{Expand any wildcards in source file paths.}
\usage{
expand_wildcards(sources, path)
}
\arguments{
\item{sources}{List of configured file paths some of which may contain
wildcards. Relative to the root of the import.}

\item{path}{Normalised path to the import directory.}
}
\value{
List of files to be used as sources
}
\description{
Expands any wildcards in the sources, prints a message if file with
name cannot be found.
}
\keyword{internal}
