% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_config.R
\name{get_default_config}
\alias{get_default_config}
\title{Create default config for a field}
\usage{
get_default_config(name)
}
\arguments{
\item{name}{The name of the field}
}
\description{
Default config is func: field_name, file: R/field_name.R
}
\keyword{internal}
