% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign_keys.R
\name{get_fk_constraints}
\alias{get_fk_constraints}
\title{Get FK constraints}
\usage{
get_fk_constraints(con)
}
\arguments{
\item{con}{Connection to get constraints for.}
}
\value{
A list representing foreign key constraints
}
\description{
Get FK constraints
}
\keyword{internal}
