% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign_keys.R
\name{get_postgres_fk}
\alias{get_postgres_fk}
\title{Get FK constraints for postgres connection}
\usage{
get_postgres_fk(con)
}
\arguments{
\item{con}{Postgres connection to get constraints for.}
}
\value{
A data frame representing foreign key constraints.
}
\description{
Get FK constraints for postgres connection
}
\keyword{internal}
