% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign_keys.R
\name{get_sqlite_fk}
\alias{get_sqlite_fk}
\title{Get FK constraints for SQLite connection}
\usage{
get_sqlite_fk(con)
}
\arguments{
\item{con}{SQLite connection to get constraints for.}
}
\value{
A data frame representing foreign key constraints.
}
\description{
Get FK constraints for SQLite connection
}
\keyword{internal}
