% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_branch}
\alias{git_branch}
\title{Get the current git branch from path.}
\usage{
git_branch(path)
}
\arguments{
\item{path}{Path to the directory to get the branch for.}
}
\value{
The git branch or error if can't be found.
}
\description{
From a path to a git controlled dir get the current active branch. This
throws an error if branch can't be retrieved.
}
\keyword{internal}
