% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_email}
\alias{git_email}
\title{Get the git user.email from git dir}
\usage{
git_email(path)
}
\arguments{
\item{path}{Path to the directory to get the git user.email for.}
}
\value{
The git user.email.
}
\description{
From a path to a git controlled dir get the user.email. This will be the
global user.email or local if it has been overriden.
}
\keyword{internal}
