% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_hash}
\alias{git_hash}
\title{Get the full hash of the current git HEAD}
\usage{
git_hash(path)
}
\arguments{
\item{path}{Path to the directory to get the hash for.}
}
\value{
Hash of HEAD.
}
\description{
Get SHA-1 hash of HEAD using \code{git rev-parse HEAD}
}
\keyword{internal}
