% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_is_clean}
\alias{git_is_clean}
\title{Check if git repo is clean}
\usage{
git_is_clean(import_dir)
}
\arguments{
\item{import_dir}{The directory of the dettl import to check.}
}
\value{
True if repo is up to date. False otherwise.
}
\description{
Check that git repo is up to date with remote by checking
\code{git status --porcelian}. If import_dir is not a git root directory
this will search parents of the import_dir for the git root
}
\keyword{internal}
