% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_repo_is_clean}
\alias{git_repo_is_clean}
\title{Check if git repo is clean}
\usage{
git_repo_is_clean(root = ".")
}
\arguments{
\item{import_path}{Path to the directory of the dettl import to check.}
}
\value{
True if repo is up to date. False otherwise.
}
\description{
Check that git repo is up to date with remote by checking
\code{git status --porcelian}. If import_path is not a git root directory
this will search parents of the import_path for the git root
}
\keyword{internal}
