% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_root_directory}
\alias{git_root_directory}
\title{Locate the git project root directory}
\usage{
git_root_directory(dir)
}
\arguments{
\item{dir}{The directory to locate the git project root from.}
}
\value{
The git project root normalized, or throws an error if this can't
be located.
}
\description{
From a subdirectory within a git project locate the root directory.
This is the directory contianing the .git files.
}
\keyword{internal}
