% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_run}
\alias{git_run}
\title{Run a git system command}
\usage{
git_run(args, root = NULL, check = FALSE)
}
\arguments{
\item{args}{The args to be passed to git}

\item{root}{The git project root directory.}

\item{check}{Whether exit code should be checked. If exist code not
successful then throw an error.}
}
\value{
Result from git including output and indication of success.
}
\description{
Check that git status using \code{git status --porcelian}. Optionally
ignoring untracked files.
}
\examples{
dettl:::git_run(c("status", "--porcelain"))
}
\keyword{internal}
