% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_status}
\alias{git_status}
\title{Check git status}
\usage{
git_status(root = NULL, ignore_untracked = FALSE)
}
\arguments{
\item{root}{The git project root directory.}

\item{ignore_untracked}{Whether untracked files should be ignored or not.
Passed as \code{--untracked-files} arg to \code{git status}}
}
\value{
Result from git including output and indication of success.
}
\description{
Check that git status using \code{git status --porcelian}. Optionally
ignoring untracked files.
}
\keyword{internal}
