% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_user}
\alias{git_user}
\title{Get the git user.name from git dir}
\usage{
git_user(path)
}
\arguments{
\item{path}{Path to the directory to get the git user.name for.}
}
\value{
The git user.name.
}
\description{
From a path to a git controlled dir get the user.name. This will be the
global user.name or local if it has been overriden.
}
\keyword{internal}
