% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_import.R
\name{ImportLog}
\alias{ImportLog}
\title{Generate log data and manage recording timings for an import}
\description{
Generate log data and manage recording timings for an import

Generate log data and manage recording timings for an import
}
\details{
Class to help help manage initialising log data, managing recording
time for log to run and importing row to the log table.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{log_data}}{The current log data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImportLog-new}{\code{ImportLog$new()}}
\item \href{#method-ImportLog-set_comment}{\code{ImportLog$set_comment()}}
\item \href{#method-ImportLog-start_timer}{\code{ImportLog$start_timer()}}
\item \href{#method-ImportLog-stop_timer}{\code{ImportLog$stop_timer()}}
\item \href{#method-ImportLog-write_log}{\code{ImportLog$write_log()}}
\item \href{#method-ImportLog-verify_first_run}{\code{ImportLog$verify_first_run()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportLog-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImportLog-new}{}}}
\subsection{Method \code{new()}}{
Build the data to be written to the import log


Create import log instance for managing creating log data and
importing to db.

Initialises log data and verifies that import can be logged. Set
a comment for the log using \code{set_comment}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportLog$new(con, log_table, path, language, mode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con}}{Connection to db to add log record to}

\item{\code{log_table}}{Name of the import log table}

\item{\code{path}}{Path to import being run}

\item{\code{language}}{The language for the import, R or sql.}

\item{\code{language}}{The language for the import, R or sql}

\item{\code{mode}}{The mode of the import, only relevant for an R import.}

\item{\code{mode}}{The mode of the import, only relevant for an R import}

\item{\code{import_path}}{Path to the import process directory.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The prepared data for the log table. This is the name
of the import, the date, comment and git information including
user name, user email, current branch and hash of HEAD.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportLog-set_comment"></a>}}
\if{latex}{\out{\hypertarget{method-ImportLog-set_comment}{}}}
\subsection{Method \code{set_comment()}}{
Set the comment on the log data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportLog$set_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{The comment to be stored with this import.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportLog-start_timer"></a>}}
\if{latex}{\out{\hypertarget{method-ImportLog-start_timer}{}}}
\subsection{Method \code{start_timer()}}{
Start the import log timer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportLog$start_timer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportLog-stop_timer"></a>}}
\if{latex}{\out{\hypertarget{method-ImportLog-stop_timer}{}}}
\subsection{Method \code{stop_timer()}}{
Stop the import log timer, adding duration to the total recorded
duration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportLog$stop_timer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportLog-write_log"></a>}}
\if{latex}{\out{\hypertarget{method-ImportLog-write_log}{}}}
\subsection{Method \code{write_log()}}{
Write the log data to the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportLog$write_log()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportLog-verify_first_run"></a>}}
\if{latex}{\out{\hypertarget{method-ImportLog-verify_first_run}{}}}
\subsection{Method \code{verify_first_run()}}{
Verify that this is the first time the import has been run.

Check whether an import with the same name has been run already. If
so then stop with a human understandable message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportLog$verify_first_run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Throws an error if an import with the same name has already been run
}
}
}
