% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into_returning.R
\name{insert_into_returning}
\alias{insert_into_returning}
\title{Insert values into table returning the primary key}
\usage{
insert_into_returning(con, table, d, key = NULL, return = NULL)
}
\arguments{
\item{con}{Connection to db to insert values into.}

\item{table}{Name of table to insert values into.}

\item{d}{Data frame of data to be inserted into db.}

\item{key}{Keys to check for already existing in the database. If trying to
add a new row with unique key constraints use this to check for existing
entries with same information.}

\item{return}{The field to return, defaults to using "id"}
}
\value{
List of returned ids from inserted values.
}
\description{
Checks whether the table already has values matching
}
\keyword{internal}
