% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_load.R
\name{insert_values_into}
\alias{insert_values_into}
\title{Insert values into table returning the primary key}
\usage{
insert_values_into(con, table, d, key = NULL, id = NULL)
}
\arguments{
\item{con}{Connection to db to insert values into.}

\item{table}{Name of table to insert values into.}

\item{d}{Data frame of data to be inserted into db.}

\item{key}{Primary key or if multiple then as composite primary key to be
checked before data is inserted into db.}

\item{id}{The field to return defaults to using "id"}
}
\value{
List of returned ids from inserted values.
}
\description{
Checks whether the table already has values matching
}
\keyword{internal}
