% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{is.PqConnection}
\alias{is.PqConnection}
\title{Tests for objects of type "PqConnection".}
\usage{
is.PqConnection(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
True if object is a "PqConnection".
}
\description{
Tests for objects of type "PqConnection".
}
\keyword{internal}
