% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{is.SQLiteConnection}
\alias{is.SQLiteConnection}
\title{Tests for objects of type "SQLiteConnection".}
\usage{
is.SQLiteConnection(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
True if object is a "SQLiteConnection".
}
\description{
Tests for objects of type "SQLiteConnection".
}
\keyword{internal}
