% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{load_sources}
\alias{load_sources}
\title{Read file fields of the dettl config yaml file.}
\usage{
load_sources(sources, path)
}
\arguments{
\item{sources}{The sources to be loaded}

\item{path}{Path to locate the sources at.}
}
\value{
Environment child of global environment which sources have been
loaded into.
}
\description{
Reads the config file, and sources any file references.
}
\keyword{internal}
