% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_dettl}
\alias{new_dettl}
\title{Create new directory and templated code for new dettl process.}
\usage{
new_dettl(name)
}
\arguments{
\item{name}{The name of the project directory to be created. Should be human
readable and meaningful. Any non a-z,0-9,_ characters will be stripped
and replaced with _s. Directory name will be preprended with created date.}
}
\description{
Create new directory and templated code for new dettl process.
}
\examples{
new_dettl("test import")
}
