% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_runner.R
\name{new_import}
\alias{new_import}
\title{Create an import object using functions defined at specified path}
\usage{
new_import(path, db_name)
}
\arguments{
\item{path}{Path to directory containing functions for import process.}

\item{db_name}{The name of the db to connect to. Connection info must be
configured via the \code{db_config.yml}.}
}
\value{
A dataImport object.
}
\description{
Create an import object using functions defined at specified path
}
