% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign_keys.R
\name{parse_constraints}
\alias{parse_constraints}
\title{Parse foreign key constraint data frame into list representation}
\usage{
parse_constraints(constraints_table)
}
\arguments{
\item{constraitns_table}{Table of constraints to parse.}
}
\value{
A list representing foreign key constraints.
}
\description{
Parse foreign key constraint data frame into list representation
}
\keyword{internal}
