% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_testing.R
\name{prepare_example_db}
\alias{prepare_example_db}
\title{Prepare example db}
\usage{
prepare_example_db(dir, add_data = FALSE, add_job_table = FALSE,
  add_log_table = TRUE)
}
\arguments{
\item{dir}{Directory to create the db in}

\item{add_data}{If TRUE data is bootstrapped to people table}

\item{add_job_table}{If TRUE also bootstrap job table related to people table}

\item{add_log_table}{If TRUE also bootstrap log table using sql in inst dir}
}
\description{
Create a SQLite database for testing called test.sqlite in specified
directory. This bootstraps a simple table called people with data optionally
added. This should only be called from a test, vignette or roxygen example.
}
\details{
Create a simple "people" table in the DB for testing the import process.
Ideally this would be in a helper-*.R file in testthat directory but we
need to be able to setup a DB for vignette.
Expect the calling function to cleanup the created db if it is not in a
tempdir.
}
\keyword{internal}
