% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_testing.R
\name{prepare_test_import}
\alias{prepare_test_import}
\title{Prepare example import inside a git repo}
\usage{
prepare_test_import(example_dir = "example",
  dettl_config = "dettl_config.yml", create_db = TRUE,
  add_data = FALSE, add_job_table = FALSE, add_log_table = TRUE)
}
\arguments{
\item{example_dir}{The example directory to copy to temp.}

\item{dettl_config}{Path to the dettl config file.}

\item{add_data}{If TRUE data is bootstrapped to people table in test DB.}

\item{add_job_table}{If TRUE also bootstrap job table related to people table.}

\item{add_log_table}{If TRUE then also bootstrap log table.}
}
\description{
Copies an example import to a new temp directory, sets up git for the
directory and creates a test SQLite DB in the temp directory as test.sqlite.
}
\details{
This should only be called from a test, vignette or roxygen example.
}
\keyword{internal}
