% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_config.R
\name{read_fields}
\alias{read_fields}
\title{Read file fields of the dettl config yaml file.}
\usage{
read_fields(fields, path)
}
\arguments{
\item{fields}{The name of the field}
}
\description{
Reads the config file, and sources any file references.
}
\keyword{internal}
