% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_config.R
\name{read_function_fields}
\alias{read_function_fields}
\title{Read file fields of the dettl config yaml file.}
\usage{
read_function_fields(fields, config, env, default_load = FALSE)
}
\arguments{
\item{fields}{The name of the field.}

\item{config}{The config being read.}

\item{env}{Environment containing functions loaded from sources.}

\item{default_load}{TRUE if this run is using default load code.}
}
\description{
Reads the config file, and sources any file references.
}
\keyword{internal}
