% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{read_r_config}
\alias{read_r_config}
\title{Get dettl config from path for an R import.}
\usage{
read_r_config(path)
}
\arguments{
\item{path}{Path to directory containing config.}
}
\value{
The config as a dettl_config object.
}
\description{
Sources any references to functions ready for use. Use a default
configuration in the result that any fields are missing. This will use the
field name as the expected file name and function name. Required fields are
extract, transform, test_queries, test and load.
}
\keyword{internal}
