% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{read_sql_config}
\alias{read_sql_config}
\title{Get dettl config from path for a SQL only import.}
\usage{
read_sql_config(path)
}
\arguments{
\item{path}{Path to directory containing config.}
}
\value{
The config as a dettl_config object.
}
\description{
Prepares load function from a file containing SQL and sources any references
to functions ready for use. Use a default configuration in the result for
any fields are missing.
}
\keyword{internal}
