% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{resolve_secrets}
\alias{resolve_secrets}
\title{Use configured vault to resolve secrets.}
\usage{
resolve_secrets(x, config)
}
\arguments{
\item{x}{List of properties some of which may be vault secrets of the form
VAULT:<path to secret>:<key to get>}

\item{config}{Cnfig containing details of the vault server to be accessed.}
}
\value{
List of properties with any secrets resolved secret.
}
\description{
This is copied from orderly.
}
\keyword{internal}
