% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_runner.R
\name{run_import}
\alias{run_import}
\title{Run the import process for a data_import object.}
\usage{
run_import(import, run_stages = NULL, dry_run = FALSE,
  comment = NULL)
}
\arguments{
\item{import}{The DataImport object.}

\item{run_stages}{Which stages should be run in the import process. Runs all
if NULL.}

\item{dry_run}{If TRUE then any database changes are rolled back when the
import completes. i.e. the load stage can be run and tests executed but the
db will be rolled back.}

\item{comment}{An optional comment to be added to the import log table for
this run.}
}
\value{
The processed DataImport object.
}
\description{
This runs the full extract, transform, test and load steps or runs specific
stages specified by 'run_stages'.
}
