% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_runner.R
\name{run_load}
\alias{run_load}
\title{Run the load step ensuring tests pass before db changes are committed.}
\usage{
run_load(con, load, transformed_data, test_queries, path, test_file,
  dry_run)
}
\arguments{
\item{con}{Connection to the database.}

\item{load}{The load function for making the DB changes.}

\item{transformed_data}{List of data frames representing the data to be
loaded to the DB.}

\item{test_queries}{Function containing queries for running on the DB before
and after the load is run. Used within the tests to check changes to the DB.}

\item{path}{Path to the import project directory.}

\item{test_file}{Path to file containing the testthat tests for verifying the
DB changes.}

\item{dry_run}{If TRUE then any database changes are rolled back when the
import completes. i.e. the load stage can be run and tests executed but the
db will be rolled back.}
}
\description{
Runs the load function on the DB within a transaction. Then run a set of
tests on the DB and rollback the changes if any should fail.
}
\keyword{internal}
