% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_runner.R
\name{run_transform}
\alias{run_transform}
\title{Run transform step.}
\usage{
run_transform(transform, extracted_data, extract_passed)
}
\arguments{
\item{transform}{Transform function loaded from config.}

\item{extracted_data}{Data returned from the extract step.}

\item{con}{The active DB connection.}

\item{path}{Path to the project directory, for locating the transform tests}

\item{transform_test}{Optional path to the transform tests. Relative to path
argument.}

\item{mode}{The type of import being run.}
}
\value{
The successfully transformed data.
}
\description{
Step is responsible for taking extracted data, running the configured
transform function on the data, checking the returned data adheres to the
db schema and running any configured user defined tests.
}
\keyword{internal}
