% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_extracted_data}
\alias{save_extracted_data}
\title{Save extracted data}
\usage{
save_extracted_data(import, file = "extracted_data.xlsx")
}
\arguments{
\item{import}{The import object to save the extracted data for.}

\item{file}{Optional file to save the data to, defaults to creating file in
current working directory called extracted_data.xlsx}
}
\description{
Saves any extracted data as separate sheet of an xlsx file.
}
\examples{
import <- new_import(system.file("examples/person_information", package = "dettl"), "test")
run_import(import, "extract")
t <- tempfile()
save_extracted_data(import, t)
}
