% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_transformed_data}
\alias{save_transformed_data}
\title{Save transformed data}
\usage{
save_transformed_data(import, file = "transformed_data.xlsx")
}
\arguments{
\item{import}{The import object to save the transformed data for.}

\item{file}{Optional file to save the data to, defaults to creating file in
current working directory called transformed_data.xlsx}
}
\description{
Saves any transformed data as separate sheet of an xlsx file.
}
