% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_testing.R
\name{setup_dettl}
\alias{setup_dettl}
\title{Setup dettl import in a tempdir.}
\usage{
setup_dettl(dir)
}
\arguments{
\item{dir}{The example directory to copy to tempdir.}
}
\description{
Setup dettl import in a tempdir.
}
\keyword{internal}
