% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sql_dialect}
\alias{sql_dialect}
\title{Get the SQL dialect used for a connection.}
\usage{
sql_dialect(con)
}
\arguments{
\item{con}{The connection to test.}
}
\value{
The dialect used, either "sqlite" or "postgres".
}
\description{
Get the SQL dialect used for a connection.
}
\keyword{internal}
