% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_import.R
\name{SqlImport}
\alias{SqlImport}
\title{Manage SQL based data import.}
\description{
Manage SQL based data import.

Manage SQL based data import.
}
\details{
This object should not be initialised directly. Use \code{\link{dettl}} to
create the object.

Import can be run by working with import object returned by
\code{\link{dettl}} or by running top-level functions. Run the import by
working with this object if you want to step through the import process
stage by stage and inspect the data after each stage.
}
\examples{
path <- dettl:::prepare_test_import(
  system.file("examples", "sql_example", package = "dettl"),
  system.file("examples", "dettl_config.yml", package = "dettl"))
import_path <- file.path(path, "sql_example")

import <- dettl::dettl(import_path, db_name = "test")
import$run_import(stage = c("extract", "transform", "load"))

}
\section{Super class}{
\code{\link[dettl:Import]{dettl::Import}} -> \code{SqlImport}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SqlImport-reload}{\code{SqlImport$reload()}}
\item \href{#method-SqlImport-read_config}{\code{SqlImport$read_config()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="begin_transaction"><a href='../../dettl/html/Import.html#method-Import-begin_transaction'><code>dettl::Import$begin_transaction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="commit_transaction"><a href='../../dettl/html/Import.html#method-Import-commit_transaction'><code>dettl::Import$commit_transaction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="extract"><a href='../../dettl/html/Import.html#method-Import-extract'><code>dettl::Import$extract()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="format"><a href='../../dettl/html/Import.html#method-Import-format'><code>dettl::Import$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="get_connection"><a href='../../dettl/html/Import.html#method-Import-get_connection'><code>dettl::Import$get_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="get_log_table"><a href='../../dettl/html/Import.html#method-Import-get_log_table'><code>dettl::Import$get_log_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="help"><a href='../../dettl/html/Import.html#method-Import-help'><code>dettl::Import$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="initialize"><a href='../../dettl/html/Import.html#method-Import-initialize'><code>dettl::Import$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="load"><a href='../../dettl/html/Import.html#method-Import-load'><code>dettl::Import$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="pre_modify_checks"><a href='../../dettl/html/Import.html#method-Import-pre_modify_checks'><code>dettl::Import$pre_modify_checks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="rollback_transaction"><a href='../../dettl/html/Import.html#method-Import-rollback_transaction'><code>dettl::Import$rollback_transaction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="run_import"><a href='../../dettl/html/Import.html#method-Import-run_import'><code>dettl::Import$run_import()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="transform"><a href='../../dettl/html/Import.html#method-Import-transform'><code>dettl::Import$transform()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqlImport-reload"></a>}}
\if{latex}{\out{\hypertarget{method-SqlImport-reload}{}}}
\subsection{Method \code{reload()}}{
Reload the objects sources to refresh source code or repair a broken
Postgres connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlImport$reload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SqlImport-read_config"></a>}}
\if{latex}{\out{\hypertarget{method-SqlImport-read_config}{}}}
\subsection{Method \code{read_config()}}{
Read and parse config from path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlImport$read_config()}\if{html}{\out{</div>}}
}

}
}
