% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlite_enable_fk}
\alias{sqlite_enable_fk}
\title{Enable foreign key constraints for SQLite connections}
\usage{
sqlite_enable_fk(con)
}
\description{
Foreign key constraints aren't enabled by default in SQLite. Ensure
they are enabled each time we connect to the db.
}
\keyword{internal}
