% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into_returning.R
\name{sqlite_execute_query}
\alias{sqlite_execute_query}
\title{Execute sql insert query returning added columns desired columns.}
\usage{
sqlite_execute_query(con, sql, table, ret, x = NULL)
}
\arguments{
\item{con}{Connection to the DB.}

\item{sql}{The sql insert query.}

\item{table}{The table to insert data into.}

\item{ret}{Vector of column names to be returned.}

\item{x}{Params for the sql query.}
}
\value{
Data frame with the newly inserted values for the specified column.
}
\description{
Execute sql insert query returning added columns desired columns.
}
\keyword{internal}
