% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into_returning.R
\name{sqlite_insert_empty_row}
\alias{sqlite_insert_empty_row}
\title{Insert an empty row and returned desired columns.}
\usage{
sqlite_insert_empty_row(con, table, return)
}
\arguments{
\item{con}{Connection to the DB.}

\item{table}{The table to add row to.}

\item{return}{Vector of column names to be returned.}
}
\value{
Data frame with the newly inserted values for the specified column.
}
\description{
Insert an empty row and returned desired columns.
}
\keyword{internal}
