% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_runner.R
\name{test_extract}
\alias{test_extract}
\title{Run tests for the extract step.}
\usage{
test_extract(con, path, extract_test, extracted_data)
}
\arguments{
\item{con}{The active DB connection.}

\item{path}{Path to the project directory, for locating the transform tests.}

\item{extract_test}{Optional path to the extract tests. Relative to path
argument.}

\item{extracted_data}{The extracted data to test.}
}
\value{
TRUE if tests pass, otherwise throws an error.
}
\description{
Runs any configured user defined tests on the extracted data. The tests are
optional for this stage and this always returns TRUE if there are no tests.
}
\keyword{internal}
