% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_runner.R
\name{test_transform}
\alias{test_transform}
\title{Run tests for transform step.}
\usage{
test_transform(
  con,
  path,
  mode,
  transform_test,
  transformed_data,
  extracted_data
)
}
\arguments{
\item{con}{The active DB connection.}

\item{path}{Path to the project directory, for locating the transform tests.}

\item{mode}{The type of import being run.}

\item{transform_test}{Optional path to the transform tests. Relative to path
argument.}

\item{transformed_data}{Data returned from the transform stage.}

\item{extracted_data}{Data returned from the extract step.}
}
\value{
TRUE is tests pass, otherwise throws an error.
}
\description{
This checks the transformed data adheres to the db schema and runs any
configured user defined tests.
}
\keyword{internal}
