% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_load.R
\name{update_child_tables}
\alias{update_child_tables}
\title{Update child tables using inserted foreign keys}
\usage{
update_child_tables(tables, table_key_pair, old_key_values, new_key_values)
}
\arguments{
\item{tables}{List of tables to be updated.}

\item{table_key_pair}{List of child tables and the foreign key field.}

\item{old_key_values}{Old values of referenced key, how foreign key is
currently identified in child table.}

\item{new_key_values}{New values of referenced key, what foreign key should
be updated to.}
}
\value{
The updated tables
}
\description{
Update child tables using inserted foreign keys
}
\keyword{internal}
