% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_runner.R
\name{verify_data}
\alias{verify_data}
\title{Verify the transformed data adhered to the DB schema.}
\usage{
verify_data(con, transformed_data)
}
\arguments{
\item{con}{The active DB connection to check the schema for.}

\item{transformed_data}{The transformed data.}
}
\description{
Transformed data returns a list of data frames. Ensure that each item of the
list matches a table in the DB and has only columns which exist in the DB.
}
\keyword{internal}
