% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_import.R
\name{verify_first_run}
\alias{verify_first_run}
\title{Verify that this is the first time the import has been run.}
\usage{
verify_first_run(con, log_table_name, log_data)
}
\arguments{
\item{con}{Connection to the DB to be imported to.}

\item{log_table_name}{The name of the log table to search for records.}

\item{log_data}{The data which will be written to the log table should
the import run.}
}
\value{
Throws an error if an import with the same name has already been run
}
\description{
Check whether an import with the same name has been run already. If
so then stop with a human understandable message.
}
\keyword{internal}
