% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{verify_table}
\alias{verify_table}
\title{Verify the data adheres to the DB schema.}
\usage{
verify_table(
  con,
  table_name,
  table,
  additional_columns = NULL,
  context_info = "",
  solution_text = ""
)
}
\arguments{
\item{con}{The active DB connection to check the schema for.}

\item{table_name}{The name of the table to check.}

\item{table}{The table to check.}

\item{additional_columns}{Vector of additional column names to check exist
in DB table.}

\item{context_info}{Info to be logged should a check fail.}

\item{solution_text}{Text describing possible solution should error occur.}
}
\description{
Verifies that
* A table with name `table_name` exists in the DB
* All the columns in `table` exist in the DB table with name `table_name`
* Any `additional_columns` exist in the DB table with name `table_name`
}
\keyword{internal}
