% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_import.R
\name{write_log}
\alias{write_log}
\title{Write the log data to the database.}
\usage{
write_log(con, log_table, log_data)
}
\arguments{
\item{con}{Connection to the database to be written to.}

\item{log_table}{The name of the table to write to.}

\item{log_data}{The log data to be written to the DB.}
}
\description{
Write the log data to the database.
}
\keyword{internal}
