% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_config.R
\name{dettl_config}
\alias{dettl_config}
\title{Get dettl config from path.}
\usage{
dettl_config(path)
}
\arguments{
\item{path}{Path to directory containing path.}
}
\value{
The config as a dettl_config object.
}
\description{
Get dettl config from path.
}
\keyword{internal}
