% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_runner.R
\name{dettl_run_transform}
\alias{dettl_run_transform}
\title{Run up to and including the transform stage of an import}
\usage{
dettl_run_transform(path, db_name = NULL)
}
\arguments{
\item{path}{Path to the import directory.}

\item{db_name}{The name of the db to connect to. Connection info must be
configured via the `dettl_config.yml`. If name is left blank this will default
to using the first db configured.}
}
\value{
The transformed data.
}
\description{
Run up to and including the transform stage of an import
}
\examples{
path <- dettl:::prepare_test_import(
  system.file("examples", "person_information", package = "dettl"),
  system.file("examples", "dettl_config.yml", package = "dettl")
)
dettl::dettl_run_transform(file.path(path, "person_information/"))
}
