% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_load.R
\name{get_default_load}
\alias{get_default_load}
\title{Get default load function.}
\usage{
get_default_load()
}
\value{
The default load function.
}
\description{
Builds the default load function using foreign key constraints read from
the database.
}
\details{
The default load function loops over the transformed data and appends each
data frame to the matching table in the database. If the appended table
contains a primary key then when the data is inserted into the database this
returns the value of the primary key for the new rows. Then loop over all
tables in which this is used as a foreign key and update the previous values
to use the returned actual values for the primary key.
}
\keyword{internal}
