% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_status}
\alias{git_status}
\title{Check git status}
\usage{
git_status(root = NULL)
}
\arguments{
\item{root}{The git project root directory.}
}
\value{
Result from git including output and indication of success.
}
\description{
Check that git status using \code{git status --porcelian}. Optionally
ignoring untracked files.
}
\keyword{internal}
