% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_import.R
\name{Import}
\alias{Import}
\title{Generic import object.}
\description{
Generic import object.

Generic import object.
}
\details{
This shouldn't be initialised directly - this object manages transactions
and the db connection and should be extended to implement actual imports.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Path to directory containing import object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Import$new()}}
\item \href{#method-reload}{\code{Import$reload()}}
\item \href{#method-pre_modify_checks}{\code{Import$pre_modify_checks()}}
\item \href{#method-format}{\code{Import$format()}}
\item \href{#method-help}{\code{Import$help()}}
\item \href{#method-get_connection}{\code{Import$get_connection()}}
\item \href{#method-get_log_table}{\code{Import$get_log_table()}}
\item \href{#method-begin_transaction}{\code{Import$begin_transaction()}}
\item \href{#method-rollback_transaction}{\code{Import$rollback_transaction()}}
\item \href{#method-commit_transaction}{\code{Import$commit_transaction()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Tidy up open connections, rolling back any active transactions


Create Import object - shouldn't be called directly. Use
\code{\link{dettl}} to create the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$new(path, db_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to directory containing import object}

\item{\code{db_name}}{Database from dettl config to create import object for}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Import` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reload"></a>}}
\if{latex}{\out{\hypertarget{method-reload}{}}}
\subsection{Method \code{reload()}}{
Reload the objects sources to refresh source code or repair a broken
Postgres connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$reload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pre_modify_checks"></a>}}
\if{latex}{\out{\hypertarget{method-pre_modify_checks}{}}}
\subsection{Method \code{pre_modify_checks()}}{
Run checks that db can be modified, this checks that:
* If require_branch set in cfg that import is for that branch
* If confirm TRUE asks users to confirm action will modify db
* If git is dirty, checks that user has explicitly said that is okay
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$pre_modify_checks(dry_run, allow_dirty_git)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dry_run}}{Whether to run in dry run mode. If TRUE then any database
changes will be rolled back. Defaults to FALSE.}

\item{\code{allow_dirty_git}}{If TRUE then skips check that the import is up to
date with remote git repo. FALSE by default.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
Custom formatter for pretty printing object summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$format(brief = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{brief}}{If TRUE then print a brief summary.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-help"></a>}}
\if{latex}{\out{\hypertarget{method-help}{}}}
\subsection{Method \code{help()}}{
Print help page for the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$help()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_connection"></a>}}
\if{latex}{\out{\hypertarget{method-get_connection}{}}}
\subsection{Method \code{get_connection()}}{
Get the database connection being used by the import. Used for testing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$get_connection()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The DBI connection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_log_table"></a>}}
\if{latex}{\out{\hypertarget{method-get_log_table}{}}}
\subsection{Method \code{get_log_table()}}{
Get the name of the log table for a particular import. This is the name
of the log table configured in the dettl config
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$get_log_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Log table name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-begin_transaction"></a>}}
\if{latex}{\out{\hypertarget{method-begin_transaction}{}}}
\subsection{Method \code{begin_transaction()}}{
Start a transaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$begin_transaction()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rollback_transaction"></a>}}
\if{latex}{\out{\hypertarget{method-rollback_transaction}{}}}
\subsection{Method \code{rollback_transaction()}}{
Rollback a transaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$rollback_transaction()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commit_transaction"></a>}}
\if{latex}{\out{\hypertarget{method-commit_transaction}{}}}
\subsection{Method \code{commit_transaction()}}{
Commit a transaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$commit_transaction()}\if{html}{\out{</div>}}
}

}
}
