% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_runner.R
\name{new_import}
\alias{new_import}
\title{Create an import object using functions defined at specified path}
\usage{
new_import(path)
}
\arguments{
\item{path}{Path to directory containing functions for import process.}
}
\value{
A dataImport object.
}
\description{
Create an import object using functions defined at specified path
}
