% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign_keys.R
\name{parse_constraints}
\alias{parse_constraints}
\title{Parse foreign key constraint data frame into list representation}
\usage{
parse_constraints(constraints_table)
}
\arguments{
\item{constraints_table}{Table of constraints to parse.}
}
\value{
A list representing foreign key constraints.
}
\description{
The representation is
}
\details{
table_name = list(
  serial = list(
    serial_field1,
    serial_field2,
    ...
  ),
  foreign = list(
    field_from_table_name = list(
      table1 = field1,
      table2 = field2,
      ...
    ),
    ...
  )
)

For a table we list the columns in that table which are serials i.e. the
columns which are generated automatically by the db. And the fields which
are referenced by a foreign key constraint. The way to read the foreign
section is \code{field_from_table_name} is a column in \code{table_name}
which is used as the referenced table in foreign key constraint from
\code{field1} in \code{table1} and \code{field2} in \code{table2}. The
implication being that \code{table_name} must exist before \code{field1} of
\code{table2} can be written to.
}
\keyword{internal}
