% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config.R
\name{read_function_fields}
\alias{read_function_fields}
\title{Read function fields of the dettl config yaml file.}
\usage{
read_function_fields(fields, config, env)
}
\arguments{
\item{fields}{List of fields with functions and whther they have to exist.}

\item{config}{The config being read.}

\item{env}{Environment containing functions loaded from sources.}
}
\description{
Reads the config file, and sources any file references.
}
\keyword{internal}
