% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_runner.R
\name{run_extract}
\alias{run_extract}
\title{Run extract step.}
\usage{
run_extract(con, extract, path)
}
\arguments{
\item{con}{The active DB connection.}

\item{extract}{Extract function loaded from config.}

\item{path}{Path to the project directory, for locating the transform tests}
}
\value{
The extracted data.
}
\description{
Step is responsible for reading data from local files and DB needed for
the transform step.
}
\keyword{internal}
