% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dettl_runner.R
\name{run_import}
\alias{run_import}
\title{Run the import process for a data_import object.}
\usage{
run_import(import, run_stages = NULL)
}
\arguments{
\item{import}{The dataImport object.}

\item{run_stages}{Which stages should be run in the import process. Runs all
if NULL.}
}
\value{
The processed dataImport object.
}
\description{
This runs the full extract, transform, test and load steps or runs specific
stages specified by 'run_stages'.
}
