% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{run_import}
\alias{run_import}
\title{Run the import process for a directory.}
\usage{
run_import(path)
}
\arguments{
\item{path}{Path to the directory.}
}
\value{
The dataImport object.
}
\description{
This runs the full extract, transform, test and load steps.
}
