% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_extracted_data}
\alias{save_extracted_data}
\title{Save extracted data}
\usage{
save_extracted_data(import, file)
}
\arguments{
\item{import}{The import object to save the extracted data for.}

\item{file}{File path at which to save the data.}
}
\description{
Saves any extracted data as separate sheet of an xlsx file.
}
\examples{
import <- dettl(system.file("examples/person_information", package = "dettl"), "test")
run_import(import, "extract")
t <- tempfile()
save_extracted_data(import, t)
}
