% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_extracted_data}
\alias{save_extracted_data}
\title{Save extracted data}
\usage{
save_extracted_data(import, file)
}
\arguments{
\item{import}{The import object to save the extracted data for.}

\item{file}{File path at which to save the data.}
}
\description{
Saves any extracted data as separate sheet of an xlsx file.
}
\examples{
path <- dettl:::prepare_test_import(
  system.file("examples", "person_information", package = "dettl"),
  system.file("examples", "dettl_config.yml", package = "dettl")
)
import <- dettl(file.path(path, "person_information"), "test")
run_import(import, "extract", dry_run = TRUE)
t <- tempfile()
save_extracted_data(import, t)
}
