% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_transformed_data}
\alias{save_transformed_data}
\title{Save transformed data}
\usage{
save_transformed_data(import, file)
}
\arguments{
\item{import}{The import object to save the transformed data for.}

\item{file}{File path at which to save the data.}
}
\description{
Saves any transformed data as separate sheet of an xlsx file.
}
\examples{
path <- prepare_test_import(
  system.file("examples", "person_information", package = "dettl"),
  system.file("examples", "dettl_config.yml", package = "dettl")
)
import <- dettl(file.path(path, "person_information"), "test")
run_import(import, c("extract", "transform"), dry_run = TRUE)
t <- tempfile()
save_transformed_data(import, t)
}
