% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_transformed_data}
\alias{save_transformed_data}
\title{Save transformed data}
\usage{
save_transformed_data(import, file)
}
\arguments{
\item{import}{The import object to save the transformed data for.}

\item{file}{File path at which to save the data.}
}
\description{
Saves any transformed data as separate sheet of an xlsx file.
}
\examples{
import <- dettl(system.file("examples/person_information", package = "dettl"), "test")
con <- import$get_connection()
people_query <- DBI::dbSendQuery(con,
  "CREATE TABLE people (
     id     BIGSERIAL PRIMARY KEY,
     name   TEXT,
     age    INTEGER,
     height INTEGER
   )"
)
DBI::dbClearResult(people_query)
run_import(import, c("extract", "transform"))
t <- tempfile()
save_transformed_data(import, t)
}
