% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_import.R
\name{SqlImport}
\alias{SqlImport}
\title{Manage SQL based data import.}
\description{
Manage SQL based data import.

Manage SQL based data import.
}
\details{
This object should not be initialised directly. Use \code{\link{dettl}} to
create the object.

Import can be run by working with import object returned by
\code{\link{dettl}} or by running top-level functions. Run the import by
working with this object if you want to step through the import process
stage by stage and inspect the data after each stage.
}
\examples{
path <- dettl:::prepare_test_import(
  system.file("examples", "sql_example", package = "dettl"),
  system.file("examples", "dettl_config.yml", package = "dettl"))
import_path <- file.path(path, "sql_example")

import <- dettl::dettl(import_path, db_name = "test")
import$run_import(stage = c("extract", "transform", "load"))

}
\section{Super class}{
\code{\link[dettl:Import]{dettl::Import}} -> \code{SqlImport}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reload}{\code{SqlImport$reload()}}
\item \href{#method-read_config}{\code{SqlImport$read_config()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="begin_transaction">}\href{../../dettl/html/Import.html#method-begin_transaction}{\code{dettl::Import$begin_transaction()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="commit_transaction">}\href{../../dettl/html/Import.html#method-commit_transaction}{\code{dettl::Import$commit_transaction()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="extract">}\href{../../dettl/html/Import.html#method-extract}{\code{dettl::Import$extract()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="format">}\href{../../dettl/html/Import.html#method-format}{\code{dettl::Import$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="get_connection">}\href{../../dettl/html/Import.html#method-get_connection}{\code{dettl::Import$get_connection()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="get_log_table">}\href{../../dettl/html/Import.html#method-get_log_table}{\code{dettl::Import$get_log_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="help">}\href{../../dettl/html/Import.html#method-help}{\code{dettl::Import$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="initialize">}\href{../../dettl/html/Import.html#method-initialize}{\code{dettl::Import$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="load">}\href{../../dettl/html/Import.html#method-load}{\code{dettl::Import$load()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="pre_modify_checks">}\href{../../dettl/html/Import.html#method-pre_modify_checks}{\code{dettl::Import$pre_modify_checks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="rollback_transaction">}\href{../../dettl/html/Import.html#method-rollback_transaction}{\code{dettl::Import$rollback_transaction()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="run_import">}\href{../../dettl/html/Import.html#method-run_import}{\code{dettl::Import$run_import()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dettl" data-topic="Import" data-id="transform">}\href{../../dettl/html/Import.html#method-transform}{\code{dettl::Import$transform()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reload"></a>}}
\if{latex}{\out{\hypertarget{method-reload}{}}}
\subsection{Method \code{reload()}}{
Reload the objects sources to refresh source code or repair a broken
Postgres connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlImport$reload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_config"></a>}}
\if{latex}{\out{\hypertarget{method-read_config}{}}}
\subsection{Method \code{read_config()}}{
Read and parse config from path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SqlImport$read_config()}\if{html}{\out{</div>}}
}

}
}
