% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{verify_table}
\alias{verify_table}
\title{Verify the data adhered to the DB schema.}
\usage{
verify_table(con, table_name, table, identical_columns = FALSE,
  context_info = "")
}
\arguments{
\item{con}{The active DB connection to check the schema for.}

\item{table_name}{The name of the table to check.}

\item{table}{The table to check.}

\item{context_info}{Info to be logged should a check fail.}
}
\description{
Ensure that table is in the DB and has only columns which exist in the DB.
And optionally check that table contains exactly the columns in the DB.
}
\keyword{internal}
